feat: query for events by file id
